+++ cards/moengage_cards_ios/lib/moengage_cards_ios.dart
@@ -40,10 +40,7 @@ class MoEngageCardsIOS extends MoEngageCardsPlatform {
   }
 
   @override
-  void setSyncCompleteListener(
-    CardsSyncListener cardsSyncListener,
-    String appId
-  ) {
+  void setSyncCompleteListener(CardsSyncListener cardsSyncListener, String appId) {
     super.setSyncCompleteListener(cardsSyncListener, appId);
     methodChannel.invokeMethod(
       methodSetCardsSyncListener,