+++ core/moengage_flutter_platform_interface/test/data_provider/json_data.dart
@@ -6,6 +6,9 @@ const String initPayload = '''
   "initConfig": {
     "pushConfig": {
       "shouldDeliverCallbackOnForegroundClick": true
+    },
+    "analyticsConfig": {
+      "shouldTrackUserAttributeBooleanAsNumber": false
     }
   }
 }