+++ cards/moengage_cards_platform_interface/CHANGELOG.md
@@ -5,6 +5,11 @@
 ## Next version
 - Refresh cards on uniqueID set
 
+# 07-08-2024
+
+## 3.0.0
+- Internal Improvements
+
 # 21-03-2024
 
 ## 2.0.0